requirements.txt
Django==5.0.2
django-tenants==3.7.0
inertia-django
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
gunicorn==22.0.0
django-cors-headers
pytz==2025.2

# Celery for background tasks
celery==5.4.0
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1

# DNS verification
dnspython==2.6.1

# HTTP requests (for Cloudflare API)
requests==2.32.3

# Excel export
openpyxl==3.1.2

# Two-Factor Authentication
pyotp==2.9.0
qrcode[pil]==7.4.2

# Supervisor for process management
supervisor==4.2.5

# Disposable email detection
disposable-email-domains==0.0.99

